LSeg/requirements.txt
paddlepaddle==2.3.2
paddlehub==2.3.0
paddlenlp==2.4.0
paddleclas==2.4.3
ftfy==6.1.1
regex==2022.9.13